+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - Old form of string still works
 - HarvestTo on HarvestStone/Forage/Shakeable/Crop now takes TinkerInventory, which puts the item into the companion's inventory (if it has one)
 - HarvestTo now works with ItemDrop abilities
+- Nop anim clips may now have a duration
 
 #### Fixed